feat: separate type and constants
