+++ src/libs/constants/gender.ts
@@ -1,0 +1,6 @@
+const gender = {
+	FEMALE: "female",
+	MALE: "male",
+};
+
+export { gender };
