+++ src/libs/constants/product-status.ts
@@ -1,0 +1,7 @@
+const productStatus = {
+	BOTH: "both",
+	FOR_RENT: "forRent",
+	FOR_SALE: "forSale",
+} as const;
+
+export { productStatus };
